+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - 33 new weapons(8 with alternate form)
 - Mirror block
 - Cryptid photo item (alchemy ingredient)
+- Skaianet Denier block that prevents Entry
+- Large cake blocks that are used to make large cake Features in Cake lands
 
 ### Changed